backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
openai==1.12.0
anthropic==0.8.1
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.26.0
pytest==7.4.3
aiohttp==3.9.3
cryptography==41.0.7
